refactor: restructure HelmChart and K8sResource nav sections
