+++ src/components/organisms/NewNavigatorPane/NavPane.tsx
@@ -3,8 +3,8 @@ import AppContext from '@src/AppContext';
 import {K8sResource} from '@models/k8sresource';
 import {HelmValuesFile} from '@models/helm';
 import KustomizationNavSection, {KustomizationNavSectionScope} from '@src/navsections/KustomizationNavSection';
-import {HelmChartNavSection, HelmChartNavSectionScope} from '@src/navsections/HelmChart.nav';
-import {K8sResourceNavSection, K8sResourceNavSectionScope} from '@src/navsections/K8sResource.nav';
+import HelmChartNavSection, {HelmChartNavSectionScope} from '@src/navsections/HelmChartNavSection';
+import K8sResourceNavSection, {K8sResourceNavSectionScope} from '@src/navsections/K8sResourceNavSection';
 import {Popover} from 'antd';
 import {PlusOutlined, FilterOutlined} from '@ant-design/icons';
 import ResourceFilter from '@components/molecules/ResourceFilter';
