+++ src/navsections/HelmChartNavSection/index.ts
@@ -1,0 +1,2 @@
+export {default} from './HelmChartNavSection';
+export type {HelmChartNavSectionScope} from './HelmChartNavSection';
