+++ src/navsections/K8sResourceNavSection/index.ts
@@ -1,0 +1,2 @@
+export {default} from './K8sResourceNavSection';
+export type {K8sResourceNavSectionScope} from './K8sResourceNavSection';
